refactor: switch review to load on bite page entry
