+++ libs/bite-tribe/api/src/lib/menu-api/menu-api.service.ts
@@ -1,0 +1,45 @@
+import { ErrorHandler, inject, Injectable } from '@angular/core';
+import { AuthService } from 'ta-firestore';
+import {
+  BehaviorSubject,
+  EMPTY,
+  from,
+  Observable,
+  skip,
+  skipWhile,
+  Subject,
+  switchMap,
+} from 'rxjs';
+import { DocumentData, FirebaseFirestore } from '@capacitor-firebase/firestore';
+import type { Menu } from 'model';
+import { AddCollectionSnapshotListenerCallbackEvent } from '@capacitor-firebase/firestore/dist/esm/definitions';
+import { getMenuById } from './utils/get-menu-by-id';
+
+export const MENU_COLLECTION = 'menus';
+
+@Injectable({ providedIn: 'root' })
+export class MenuApiService {
+  private readonly authService = inject(AuthService);
+  private readonly errorHandler = inject(ErrorHandler);
+
+  loadMenu(menuId: string): Promise<Menu | undefined> {
+    return getMenuById(menuId);
+  }
+
+  handleError(err: any): typeof EMPTY {
+    console.error('Error fetching menu:', err);
+    this.errorHandler.handleError(err);
+    return EMPTY;
+  }
+
+  async saveMenu(menu: Menu): Promise<void> {
+    await FirebaseFirestore.updateDocument({
+      reference: `${MENU_COLLECTION}/${menu.id}`,
+      data: {
+        categories: menu.categories,
+        updatedAt: new Date().toISOString(),
+        updatedAtTimestamp: Date.now(), // numeric timestamp for easier queries
+      },
+    });
+  }
+}
