+++ libs/bite-tribe/api/src/lib/restaurant-api/restaurant-api.service.ts
@@ -2,7 +2,7 @@ import { inject, Injectable } from '@angular/core';
 import { AuthService } from 'ta-firestore';
 import { FirebaseFirestore } from '@capacitor-firebase/firestore';
 import type { Link, Restaurant } from 'model';
-import { MENU_COLLECTION } from '../menu-api.service';
+import { MENU_COLLECTION } from '../menu-api/menu-api.service';
 import { BITE_COLLECTION, RESTAURANT_COLLECTION } from '../utils/constants';
 import { getRestaurantById } from './utils/load-restaurant-by-id';
