+++ libs/bite-tribe/api/src/lib/utils/constants.ts
@@ -1,5 +1,6 @@
 export const BITE_COLLECTION = 'bites';
 export const BUCKETLIST_COLLECTION = 'bucketlists';
 export const RESTAURANT_COLLECTION = 'restaurants';
+export const REVIEW_COLLECTION = 'reviews';
 
 export const DEFAULT_SEARCH_RADIUS_IN_M = 15 * 1000;
