+++ libs/bite-tribe/store/src/lib/menus/effects.ts
@@ -2,11 +2,12 @@ import { inject, Injectable } from '@angular/core';
 import { Actions, createEffect, ofType } from '@ngrx/effects';
 import { BiteTribeApiService } from 'bite-tribe/api';
 import { routerNavigatedAction } from '@ngrx/router-store';
-import { map, skipWhile, switchMap, tap } from 'rxjs';
+import { filter, from, map, switchMap, tap } from 'rxjs';
 import { Store } from '@ngrx/store';
 import { menuId } from '../router/selectors';
 import { MenuActions } from './actions';
 import { fromAuth } from 'ta-firestore';
+import { toSignal } from '@angular/core/rxjs-interop';
 
 @Injectable()
 export class MenuEffects {
@@ -14,23 +15,18 @@ export class MenuEffects {
   private readonly store = inject(Store);
   private readonly api = inject(BiteTribeApiService);
 
-  startListener$ = createEffect(() => {
-    return this.actions$.pipe(
-      ofType(fromAuth.AuthActions.loginSucceeded),
-      switchMap(() => this.api.menus$()),
-      map((menus) => MenuActions.loadedMenusFromAPI({ menus })),
-    );
-  });
+  menuId = toSignal(this.store.select(menuId));
 
   loadMenuFromApi$ = createEffect(() => {
     return this.actions$.pipe(
       ofType(routerNavigatedAction),
-      switchMap(() =>
-        this.store.select(menuId).pipe(
-          skipWhile((menuId) => !menuId),
-          switchMap((menuId) => {
-            return this.api.loadMenu(menuId);
-          }),
+      filter(({ payload }) => {
+        return payload.event.urlAfterRedirects.includes(`/menu/`);
+      }),
+      switchMap(() => {
+        const menuId = this.menuId();
+
+        return from(this.api.loadMenu(menuId)).pipe(
           map((menu) => {
             if (!menu) {
               return MenuActions.noMenuFound();
@@ -38,8 +34,8 @@ export class MenuEffects {
 
             return MenuActions.loadedMenuFromAPI({ menu });
           }),
-        ),
-      ),
+        );
+      }),
     );
   });
